+++ Application/Service/Empolyee/IhousingService.cs
@@ -10,7 +10,7 @@ public interface IHousingService
 {
     Task<Result<IEnumerable<HousingResponse>>> GetAllEmployee();
     Task<Result<IEnumerable<HousingResponse>>> Get(string Name);
-    Task<Result<IEnumerable<HousingResponse>>> GetWithManagerIqama(int ManagerIqamaNo);
+    Task<Result<IEnumerable<HousingResponse>>> GetWithManagerIqama(long ManagerIqamaNo);
     Task<Result<HousingResponse>> CreateAsync(HousingRequest Request);
     Task<Result<UHousingResponse>> UpdateAsync(HousingRequest  Request);
     Task<Result> DeleteAsync(string Name, CancellationToken cancellationToken = default);
